+++ CHANGELOG.md
@@ -30,6 +30,10 @@ You can check your current version with the following command:
 
 For more information, see [UP42 Python package description](https://pypi.org/project/up42-py/).
 
+### June 27, 2025 [minor]
+- Added `FeasibilityStudy` active record class to `tasking` module.
+- Exported `FeasibilityStudy` and `FeasibilityStudySorting` to `up42` namespace.
+
 ## 2.3.1
 **June 26, 2025**